requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
spotipy==2.23.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
apscheduler==3.10.4
jinja2==3.1.2
python-dotenv==1.0.0